updated DeleteColumnService

+++ app/services/columns/delete_column_service.rb
@@ -4,17 +4,18 @@ module Columns
   class DeleteColumnService
     include Callable
 
-    attr_reader :user_id, :id
+    attr_reader :current_user, :id
 
-    def initialize(user_id, id)
-      @user_id = user_id
+    def initialize(current_user, id)
+      @current_user = current_user
       @id = id
     end
 
     def call
-      column = Column.find_by!(user_id: @user_id, id: @id)
+      raise Exceptions::Unauthorized, 'Unauthorized' unless @current_user.present?
+      column = Column.find_by!(id: @id)
+      raise Exceptions::Forbidden, 'Forbidden' unless column.user == @current_user
       column.destroy
-
       OpenStruct.new(success: column.destroyed?, errors: column.errors)
     end
   end
